+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 
 #### 🚀 Updates
 
+- Added support for updating the `~/.prototools` file (root of user home directory).
+  - Added `--pin=user` to `proto install`.
+  - Added `--user` to `proto alias`, `unalias`, `pin`, `unpin`, `plugin add`, and `plugin remove`.
+  - Added `store` as an alias for `global`.
+  - Added `cwd` as an alias for `local`.
+  - Added `home` as an alias for `user`.
 - Added new `settings.offline` settings that control how offline checks work.
 
 #### ⚙️ Internal